Add changes Co-authored by John Purvis

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/AccountDao.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface AccountDao {
 
-    Account getAccount(int accountId);
+    Account getAccount(long accountId);
 
     List<Account> getAllAccounts();
